# Conflicts:
#	AGENTS.md
#	electron/src/__tests__/fileExplorer.test.ts
#	mobile/src/screens/AssetViewerScreen.zoom.test.tsx
#	packages/agents/src/app-build/judge.ts
#	packages/agents/src/capabilities/models.ts
#	packages/agents/src/capabilities/shared.ts
#	packages/agents/src/capabilities/sketches.ts
#	packages/agents/src/capabilities/storyboards.ts
#	packages/agents/src/capabilities/timelines.ts
#	packages/agents/src/evals/codeact-api-core.ts
#	packages/agents/src/evals/goal-judge.ts
#	packages/agents/src/host-binaries.ts
#	packages/agents/src/long-term-memory.ts
#	packages/agents/src/sandbox-media.ts
#	packages/agents/src/security-monitor.ts
#	packages/agents/src/subagent.ts
#	packages/agents/src/task-planner.ts
#	packages/agents/src/tools/base-tool.ts
#	packages/agents/src/tools/plan-builder-tools.ts
#	packages/agents/src/tools/submit-code-tool.ts
#	packages/app-runtime/src/bundle.ts
#	packages/audio-nodes/src/nodes/output.ts
#	packages/cli/src/commands/agent.ts
#	packages/cli/src/nodetool.ts
#	packages/kernel/src/graph-utils.ts
#	packages/kernel/src/graph.ts
#	packages/kernel/src/supervisor.ts
#	packages/llm-nodes/src/nodes/director.ts
#	packages/llm-nodes/src/nodes/openai.ts
#	packages/llm-nodes/src/nodes/shots.ts
#	packages/node-sdk/src/code-body.ts
#	packages/node-sdk/src/docs/workflows.ts
#	packages/text-nodes/src/nodes/lib-svg.ts
#	packages/video-nodes/src/nodes/ffmpeg-helpers.ts
#	packages/websocket/src/fake-runtime.ts
#	packages/websocket/src/mcp-agent-tools.ts
#	packages/websocket/src/server.ts
#	packages/websocket/src/triggers/registration-sync.ts
#	packages/websocket/src/trpc/routers/threads.ts
#	packages/websocket/src/ws-adapter.ts
#	web/src/components/chat/hooks/useDragAndDrop.ts
#	web/src/components/execution/ExecutionTree.tsx
#	web/src/components/node_types/CodeBody.tsx
#	web/src/components/node_types/code_assistant/CodeAssistantDialog.tsx
#	web/src/components/node_types/editing/ExtractVideoFrameBody.tsx
#	web/src/components/node_types/editing/GetVariableBody.tsx
#	web/src/components/node_types/editing/PainterBody.tsx
#	web/src/components/node_types/editing/PromptComposerBody.tsx
#	web/src/components/sketch/SketchCanvasResizeHandles.tsx
#	web/src/components/ui_primitives/Popover.tsx
#	web/src/lib/tools/builtin/puck.ts
#	web/src/stores/workflowUpdates.ts
#	web/src/utils/imageUploadValidation.ts

+++ packages/transformers-js-nodes/src/nodes/text-to-speech.ts
@@ -115,7 +115,7 @@ export class TextToSpeechNode extends BaseNode {
       const tts = await getKokoro(repoId!, dtype, device);
       const voice = asString(this.voice) || "af_heart";
       const result = await tts.generate(text, { voice: voice as never });
-      samples = result.audio as Float32Array;
+      samples = result.audio;
       samplingRate = result.sampling_rate ?? samplingRate;
     } else {
       const pipeline = await getPipeline<
